+++ NEWS.md
@@ -4,7 +4,7 @@
 - DART installer and updater (pytools4dart.dart.install & pytools4dart.dart.update)
 
 ## Change
-- deprecated OBJtools get_gnames, get_dims, get_center for objreader properties  
+- deprecated OBJtools read, get_gnames, get_dims, get_center for objreader properties  
 
 # 1.1.6